upadate the type of the y for tensorflow

+++ submissions/deep_learn/estimator.py
@@ -44,8 +44,9 @@ def _dice_coefficient_loss(y_true, y_pred):
 
 
 def _dice_coefficient(y_true, y_pred, smooth=1.):
-    y_true_f = K.flatten(y_true)
-    y_pred_f = K.flatten(y_pred)
+    y_true_f = tf.cast(K.flatten(y_true), tf.float64)
+    y_pred_f = tf.cast(K.flatten(y_pred), tf.float64)
+
     intersection = K.sum(y_true_f * y_pred_f)
 
     return ((2. * intersection + smooth) / (K.sum(y_true_f) +
